# packages/mtmai/mtmai-0.3.1016.tar.gz/mtmai-0.3.1016/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.1016"
  },
  "paths": {
...
    "/api/v1/chat/completions": {
      "post": {
        "tags": [
          "openai"
        ],
        "summary": "Chat Completions",
        "description": "\u4ee5\u517c\u5bb9 openai completions \u534f\u8bae\u7684\u65b9\u5f0f \u53cd\u4ee3\u5176\u4ed6 ai \u63d0\u4f9b\u5546\u7684 completions api",
        "operationId": "openai-chat_completions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}